+++ src/core/store-id.ts
@@ -1,6 +1,7 @@
 import { storagesStore } from '@src/infrastructure/prun-api/data/storage';
 import { sitesStore } from '@src/infrastructure/prun-api/data/sites';
 import { warehousesStore } from '@src/infrastructure/prun-api/data/warehouses';
+import { shipsStore } from '@src/infrastructure/prun-api/data/ships';
 
 export function getInvStoreId(invParameter: string) {
   invParameter = invParameter.toLowerCase().trim();
@@ -13,6 +14,10 @@ export function getInvStoreId(invParameter: string) {
   if (!store) {
     const warehouse = warehousesStore.getByEntityNaturalId(invParameter);
     store = stores.find(x => x.id === warehouse?.storeId);
+  }
+  if (!store) {
+    const ship = shipsStore.getByRegistration(invParameter);
+    store = stores.find(x => x.id === ship?.idShipStore);
   }
   return store;
 }